Remove ProcessLifecycleOwnerInitializer workaround from manifest
The androidx.lifecycle.ProcessLifecycleOwnerInitializer provider
was added to the manifest as a workaround for incorrect manifest
merging in b/112587252.  Soong has been fixed to only merge the
manifests at the final APK, preserving the right value of
${applicationId}, so the workaround is no longer necessary.

An upcoming drop of androidx to the platform removes the
androidx.lifecycle.ProcessLifecycleOwnerInitializer in favor
of androidx.startup.InitializationProvider, which causes failures
at runtime if the workaround is not removed.

Bug: 197536008
Test: ProcessLifecycleOwnerInitializer still in compiled AndroidManifest.xml
Change-Id: I3094beacd63b1a87b8689b66fd744aeb6a037e68
Merged-In: I5690b23933cb074acde6d4bc08f4d656f5c7db91